+++ modules/apps/frontend-taglib/frontend-taglib-clay-sample-web/src/main/java/com/liferay/frontend/taglib/clay/sample/web/constants/ClaySamplePortletKeys.java
@@ -24,9 +24,6 @@ public class ClaySamplePortletKeys {
 	public static final String CLAY_SAMPLE =
 		"com_liferay_clay_sample_web_portlet_ClaySamplePortlet";
 
-	public static final String CLAY_SAMPLE_TABLE_DISPLAY_CONTEXT =
-		"CLAY_SAMPLE_TABLE_DISPLAY_CONTEXT";
-
 	public static final String DROPDOWNS_DISPLAY_CONTEXT =
 		"DROPDOWNS_DISPLAY_CONTEXT";
 
@@ -35,5 +32,8 @@ public class ClaySamplePortletKeys {
 
 	public static final String NAVIGATION_BARS_DISPLAY_CONTEXT =
 		"NAVIGATION_BARS_DISPLAY_CONTEXT";
+
+	public static final String TABLES_DISPLAY_CONTEXT =
+		"TABLES_DISPLAY_CONTEXT";
 
 }
