+++ modules/apps/frontend-taglib/frontend-taglib-clay-sample-web/src/main/resources/META-INF/resources/partials/tables.jsp
@@ -19,5 +19,5 @@
 <blockquote><p>Tables are lists of properties of items arranged as rows and columns. Each table cell can use a custom renderer so that its data can be displayed in the best way.</p></blockquote>
 
 <clay:table
-	tableDisplayContext="<%= claySampleTableDisplayContext %>"
+	displayContext="<%= tablesDisplayContext %>"
 />
